API/requirements.txt
async-timeout==4.0.2
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
python-dotenv==1.0.0
redis==4.5.1
six==1.16.0
Werkzeug==2.2.3
zope.event==4.6
zope.interface==5.5.2
